Refactor `FixedProductMarketMakerCreation` events
Module `FixedProductMarketMakerCreation` watch event from
`FixedProductionMarketMakerFactory` contract to `marketEvents.js` file.
